+++ src/tree/registries/IRegistryProviderTreeItem.ts
@@ -1,0 +1,10 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { ICachedRegistryProvider } from "./ICachedRegistryProvider";
+
+export interface IRegistryProviderTreeItem {
+    cachedProvider: ICachedRegistryProvider;
+}
