+++ src/tree/registries/dockerV2/GenericDockerV2RegistryTreeItem.ts
@@ -7,7 +7,8 @@ import { RequestPromiseOptions } from "request-promise-native";
 import { AzExtParentTreeItem, AzExtTreeItem, IActionContext, parseError } from "vscode-azureextensionui";
 import { nonNullProp } from "../../../utils/nonNull";
 import { registryRequest } from "../../../utils/registryRequestUtils";
-import { ICachedRegistryProvider, IRegistryProviderTreeItem } from "../IRegistryProvider";
+import { ICachedRegistryProvider } from "../ICachedRegistryProvider";
+import { IRegistryProviderTreeItem } from "../IRegistryProviderTreeItem";
 import { getRegistryContextValue, registryProviderSuffix, registrySuffix } from "../registryContextValues";
 import { getRegistryPassword } from "../registryPasswords";
 import { IDockerCliCredentials } from "../RegistryTreeItemBase";
